# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 노들섬 홍보 통합 시스템 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List organizations (admin sees all, user sees own)
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationInput"
      responses:
        "201":
          description: Created organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get organization detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"

  # === EVENTS ===
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: organizationId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventList"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event detail with full workflow data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/timeline:
    get:
      operationId: getEventTimeline
      tags: [events]
      summary: Get full event timeline (all activity log)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event timeline entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEntry"

  # === PROMOTION ZONES ===
...
components:
  schemas:
...
    Organization:
      type: object
      required: [id, name, contactEmail, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        createdAt:
          type: string
    OrganizationInput:
      type: object
      required: [name, contactEmail]
      properties:
        name:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
    Event:
      type: object
      required: [id, title, status, organizationId, startDate, endDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, submitted, approved, revision_requested, rejected, completed]
        organizationId:
          type: integer
        organizationName:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        venue:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EventList:
      type: object
      required: [events, total, page, limit]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EventDetail:
      type: object
      required: [id, title, status, organizationId, startDate, endDate, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, submitted, approved, revision_requested, rejected, completed]
        organizationId:
          type: integer
        organizationName:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        venue:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        adminNote:
          type: ["string", "null"]
        promotionRequests:
          type: array
          items:
            $ref: "#/components/schemas/PromotionRequest"
        assets:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        schedules:
          type: array
          items:
            $ref: "#/components/schemas/Schedule"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        emailLogs:
          type: array
          items:
            $ref: "#/components/schemas/EmailLog"
        createdAt:
          type: string
        updatedAt:
          type: string
    EventInput:
      type: object
      required: [title, startDate, endDate]
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        venue:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        tags:
          type: array
          items:
            type: string
    EventUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        venue:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        tags:
          type: array
          items:
            type: string
        adminNote:
          type: string
        status:
          type: string
          enum: [draft, submitted, approved, revision_requested, rejected, completed]
    TimelineEntry:
      type: object
      required: [id, eventId, type, content, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        type:
          type: string
          enum: [comment, approval, rejection, revision_request, asset_upload, status_change, email_sent, schedule_created]
        content:
          type: string
        authorName:
          type: ["string", "null"]
        authorRole:
          type: ["string", "null"]
        createdAt:
          type: string
...
    PromotionRequest:
      type: object
      required: [id, eventId, zoneId, status, requestedStartDate, requestedEndDate, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        zoneId:
          type: integer
        zoneName:
          type: ["string", "null"]
        zoneType:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, revision_requested, final_approved]
        requestedStartDate:
          type: string
        requestedEndDate:
          type: string
        notes:
          type: ["string", "null"]
        adminComment:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Asset:
      type: object
      required: [id, eventId, name, currentVersion, totalVersions, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        zoneId:
          type: ["integer", "null"]
        zoneName:
          type: ["string", "null"]
        currentVersion:
          type: integer
        totalVersions:
          type: integer
        selectedVersionId:
          type: ["integer", "null"]
        latestVersionUrl:
          type: ["string", "null"]
        latestVersionFileType:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending_review, revision_requested, approved, final_selected]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Schedule:
      type: object
      required: [id, eventId, zoneId, startDate, endDate, status, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        eventTitle:
          type: ["string", "null"]
        zoneId:
          type: integer
        zoneName:
          type: ["string", "null"]
        zoneType:
          type: ["string", "null"]
        zoneColor:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [scheduled, posted, completed, cancelled]
        assetVersionId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Comment:
      type: object
      required: [id, eventId, content, authorName, authorRole, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        content:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
          enum: [user, admin, super_admin]
        isAdminOnly:
          type: boolean
        createdAt:
          type: string
...
    EmailLog:
      type: object
      required: [id, eventId, emailType, recipientEmail, status, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        emailType:
          type: string
          enum: [application_received, approved, revision_requested, final_approved, posting_complete, custom]
        recipientEmail:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [sent, failed, pending]
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string
...