# lib/api-zod/src/generated/types/organizationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 노들섬 홍보 통합 시스템 API
 * OpenAPI spec version: 0.1.0
 */

export interface OrganizationInput {
  name: string;
  contactName?: string;
  contactEmail: string;
  contactPhone?: string;
}